+++ testing/testing_zhemm_vbatched.cpp
@@ -289,15 +289,15 @@ int main( int argc, char** argv)
                 bool okay = (magma_error < tol);
                 status += ! okay;
 
-                    printf("  %-10d  %-5d %-5d %-7.2f ( %-7.2f )  %-7.2f ( %-7.2f )  %-8.2e  %s\n",
-                       (int) batchCount, (int) max_M, (int) max_N,
+                    printf("  %-10lld  %-5lld %-5lld %-7.2f ( %-7.2f )  %-7.2f ( %-7.2f )  %-8.2e  %s\n",
+                       (long long)batchCount, (long long)max_M, (long long)max_N,
                        magma_perf,  1000.*magma_time,
                        cpu_perf,    1000.*cpu_time,
                        magma_error, (okay ? "ok" : "failed"));
             }
             else {
-                    printf("  %-10d  %-5d %-5d %-7.2f ( %-7.2f )  ------- ( ------- )  --------  \n",
-                       (int) batchCount, (int) max_M, (int) max_N,
+                    printf("  %-10lld  %-5lld %-5lld %-7.2f ( %-7.2f )  ------- ( ------- )  --------  \n",
+                       (long long)batchCount, (long long)max_M, (long long)max_N,
                        magma_perf,  1000.*magma_time);
             }
